wikipedia/ZS-BERT/requirements.txt
torchvision
transformers
sentence-transformers
scikit-learn
pandas
numpy
html5lib
beautifulsoup4